Initial commit of fake client for network testing
This was already pushed to 2015-CS454/dd-team#test_client but makes more sense
as a standalone repository due to changes with the ConnectionManager as well as
the network classes which were customized for testing purposes.

The initial specification for the protocol has some glaring inconsistencies.
Due to this, we either hardcode some values or entirely skip testing them until
the protocol can be refined.